+++ test/sample.jl
@@ -96,6 +96,30 @@ end
 
     @time println(prob(ft, top, env))
     @time println(prob(ft, top, env))
+    @time println(grad(ft, top, env))
+    @time println(smeas(ft, top))
+end
+
+@testset "example" begin
+    ft = FTree()
+
+    @basic A
+    @repeated B, C
+    
+    top = (A | B) & C
+    
+    env = @parameters begin
+        A = 0.9
+        B = 0.98
+        C = 0.89
+    end
+    
+    println(prob(ft, top, env))
+    println(mcs(ft, top))
+    println(smeas(ft, top))
+    println(bmeas(ft, top, env))
+    println(c1meas(ft, top, env))
+    println(c0meas(ft, top, env))
 end
 
 # @testset "Sample2" begin